src/main/java/com/example/service/impl/SellerServiceImpl.java
package com.example.service.impl;

import com.example.entity.SellerInfo;
import com.example.repository.SellerInfoRepository;
import com.example.service.SellerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Created by qidd on 2018-3-25
 */
@Service
public class SellerServiceImpl implements SellerService {
    @Autowired
    private SellerInfoRepository repository;


    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {

        return repository.findByOpenid(openid);

    }
}
